# backend/docs/swagger.yaml
basePath: /api/v1
definitions:
...
  templatev25_internal_http_dto.ErrorResponse:
    description: Error response structure
    properties:
      code:
        description: Error code
        example: BAD_REQUEST
        type: string
      details:
        additionalProperties:
          type: string
        description: Detailed error information
        type: object
      message:
        description: Error message
        example: Invalid request parameters
        type: string
      request_id:
        description: Request ID for debugging
        example: 550e8400-e29b-41d4-a716-446655440000
        type: string
    type: object
...
  templatev25_internal_http_dto.LoginRequest:
    properties:
      email:
        type: string
      password:
        minLength: 8
        type: string
    required:
    - email
    - password
    type: object
  templatev25_internal_http_dto.LoginResponse:
    properties:
      access_token:
        type: string
      expires_at:
        type: integer
      mfa_token:
        type: string
      requires_mfa:
        type: boolean
      user:
        $ref: '#/definitions/templatev25_internal_http_dto.UserInfo'
    type: object
...
  templatev25_internal_http_dto.Response:
    description: Standard API response wrapper
    properties:
      code:
        description: Business logic code (OK, BAD_REQUEST, etc.)
        example: OK
        type: string
      data:
        description: Response data (can be any type)
      details:
        description: Error details for validation errors
      message:
        description: Human-readable message
        example: success
        type: string
      request_id:
        description: Unique request identifier
        example: 550e8400-e29b-41d4-a716-446655440000
        type: string
    type: object
...
  templatev25_internal_http_dto.UserInfo:
    properties:
      email:
        type: string
      first_name:
        type: string
      id:
        type: integer
      last_name:
        type: string
      status:
        type: string
    type: object
...
  templatev25_internal_http_dto.VerifyBackupCodeRequest:
    properties:
      code:
        type: string
      mfa_token:
        type: string
    required:
    - code
    - mfa_token
    type: object
  templatev25_internal_http_dto.VerifyMFARequest:
    properties:
      code:
        type: string
      mfa_token:
        type: string
    required:
    - code
    - mfa_token
    type: object
info:
  contact: {}
  description: Fiber + Gorm + Auth + RBAC demo
  title: Template Backend API
  version: "1.0"
paths:
...
  /auth/local/login:
    post:
      consumes:
      - application/json
      description: Authenticate user with email and password. Returns MFA token if
        MFA is enabled.
      parameters:
      - description: Login credentials
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/templatev25_internal_http_dto.LoginRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/templatev25_internal_http_dto.LoginResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/templatev25_internal_http_dto.ErrorResponse'
        "401":
          description: Invalid credentials
          schema:
            $ref: '#/definitions/templatev25_internal_http_dto.ErrorResponse'
        "423":
          description: Account locked
          schema:
            $ref: '#/definitions/templatev25_internal_http_dto.ErrorResponse'
      summary: Login with email and password
      tags:
      - local-auth
  /auth/local/logout:
    post:
      description: Logout and invalidate current session
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/templatev25_internal_http_dto.Response'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/templatev25_internal_http_dto.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Logout current session
      tags:
      - local-auth
  /auth/local/logout-all:
    post:
      description: Logout and invalidate all user sessions
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/templatev25_internal_http_dto.Response'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/templatev25_internal_http_dto.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Logout all sessions
      tags:
      - local-auth
...
  /auth/local/refresh:
    post:
      description: Extend session expiration time
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/templatev25_internal_http_dto.LoginResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/templatev25_internal_http_dto.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Refresh session
      tags:
      - local-auth
  /auth/local/verify-backup-code:
    post:
      consumes:
      - application/json
      description: Verify backup code to complete login when TOTP is unavailable
      parameters:
      - description: Backup code verification
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/templatev25_internal_http_dto.VerifyBackupCodeRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/templatev25_internal_http_dto.LoginResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/templatev25_internal_http_dto.ErrorResponse'
        "401":
          description: Invalid code
          schema:
            $ref: '#/definitions/templatev25_internal_http_dto.ErrorResponse'
      summary: Verify backup code
      tags:
      - local-auth
  /auth/local/verify-mfa:
    post:
      consumes:
      - application/json
      description: Verify TOTP code to complete login
      parameters:
      - description: MFA verification
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/templatev25_internal_http_dto.VerifyMFARequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/templatev25_internal_http_dto.LoginResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/templatev25_internal_http_dto.ErrorResponse'
        "401":
          description: Invalid code
          schema:
            $ref: '#/definitions/templatev25_internal_http_dto.ErrorResponse'
      summary: Verify MFA code
      tags:
      - local-auth
...
schemes:
- http
- https
securityDefinitions:
  BearerAuth:
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"